@@ -1,117 +1,154 @@
-import Vue from 'vue';
-import Router from 'vue-router';
+import {createRouter, createWebHistory} from "vue-router";
+import Home from "../views/Home.vue";
 
-Vue.use(Router);
+const routes = [
+    {
+        path: '/',
+        redirect: '/dashboard'
+    }, {
+        path: "/",
+        name: "Home",
+        component: Home,
+        children: [
+            {
+                path: "/dashboard",
+                name: "Dashboard",
+                meta: {
+                    title: '系统首页'
+                },
+                component: () => import (
+                /* webpackChunkName: "dashboard" */
+                "../views/Dashboard.vue")
+            }, {
+                path: "/table",
+                name: "BaseTable",
+                meta: {
+                    title: '表格'
+                },
+                component: () => import (
+                /* webpackChunkName: "table" */
+                "../views/BaseTable.vue")
+            }, {
+                path: "/charts",
+                name: "BaseCharts",
+                meta: {
+                    title: '图表'
+                },
+                component: () => import (
+                /* webpackChunkName: "charts" */
+                "../views/BaseCharts.vue")
+            }, {
+                path: "/form",
+                name: "BaseForm",
+                meta: {
+                    title: '表单'
+                },
+                component: () => import (
+                /* webpackChunkName: "form" */
+                "../views/BaseForm.vue")
+            }, {
+                path: "/tabs",
+                name: "Tabs",
+                meta: {
+                    title: 'tab标签'
+                },
+                component: () => import (
+                /* webpackChunkName: "tabs" */
+                "../views/Tabs.vue")
+            }, {
+                path: "/donate",
+                name: "Donate",
+                meta: {
+                    title: '鼓励作者'
+                },
+                component: () => import (
+                /* webpackChunkName: "donate" */
+                "../views/Donate.vue")
+            }, {
+                path: "/permission",
+                name: "Permission",
+                meta: {
+                    title: '权限管理',
+                    permission: true
+                },
+                component: () => import (
+                /* webpackChunkName: "permission" */
+                "../views/Permission.vue")
+            }, {
+                path: "/i18n",
+                name: "I18n",
+                meta: {
+                    title: '国际化语言'
+                },
+                component: () => import (
+                /* webpackChunkName: "i18n" */
+                "../views/I18n.vue")
+            }, {
+                path: "/upload",
+                name: "Upload",
+                meta: {
+                    title: '上传插件'
+                },
+                component: () => import (
+                /* webpackChunkName: "upload" */
+                "../views/Upload.vue")
+            }, {
+                path: "/icon",
+                name: "Icon",
+                meta: {
+                    title: '自定义图标'
+                },
+                component: () => import (
+                /* webpackChunkName: "icon" */
+                "../views/Icon.vue")
+            }, {
+                path: '/404',
+                name: '404',
+                meta: {
+                    title: '找不到页面'
+                },
+                component: () => import (/* webpackChunkName: "404" */
+                '../views/404.vue')
+            }, {
+                path: '/403',
+                name: '403',
+                meta: {
+                    title: '没有权限'
+                },
+                component: () => import (/* webpackChunkName: "403" */
+                '../views/403.vue')
+            }
+        ]
+    }, {
+        path: "/login",
+        name: "Login",
+        meta: {
+            title: '登录'
+        },
+        component: () => import (
+        /* webpackChunkName: "login" */
+        "../views/Login.vue")
+    }
+];
 
-export default new Router({
-    routes: [
-        {
-            path: '/',
-            redirect: '/dashboard'
-        },
-        {
-            path: '/',
-            component: () => import(/* webpackChunkName: "home" */ '../components/common/Home.vue'),
-            meta: { title: '自述文件' },
-            children: [
-                {
-                    path: '/dashboard',
-                    component: () => import(/* webpackChunkName: "dashboard" */ '../components/page/Dashboard.vue'),
-                    meta: { title: '系统首页' }
-                },
-                {
-                    path: '/icon',
-                    component: () => import(/* webpackChunkName: "icon" */ '../components/page/Icon.vue'),
-                    meta: { title: '自定义图标' }
-                },
-                {
-                    path: '/table',
-                    component: () => import(/* webpackChunkName: "table" */ '../components/page/BaseTable.vue'),
-                    meta: { title: '基础表格' }
-                },
-                {
-                    path: '/tabs',
-                    component: () => import(/* webpackChunkName: "tabs" */ '../components/page/Tabs.vue'),
-                    meta: { title: 'tab选项卡' }
-                },
-                {
-                    path: '/form',
-                    component: () => import(/* webpackChunkName: "form" */ '../components/page/BaseForm.vue'),
-                    meta: { title: '基本表单' }
-                },
-                {
-                    // 富文本编辑器组件
-                    path: '/editor',
-                    component: () => import(/* webpackChunkName: "editor" */ '../components/page/VueEditor.vue'),
-                    meta: { title: '富文本编辑器' }
-                },
-                {
-                    // markdown组件
-                    path: '/markdown',
-                    component: () => import(/* webpackChunkName: "markdown" */ '../components/page/Markdown.vue'),
-                    meta: { title: 'markdown编辑器' }
-                },
-                {
-                    // 图片上传组件
-                    path: '/upload',
-                    component: () => import(/* webpackChunkName: "upload" */ '../components/page/Upload.vue'),
-                    meta: { title: '文件上传' }
-                },
-                {
-                    // vue-schart组件
-                    path: '/charts',
-                    component: () => import(/* webpackChunkName: "chart" */ '../components/page/BaseCharts.vue'),
-                    meta: { title: 'schart图表' }
-                },
-                {
-                    // 拖拽列表组件
-                    path: '/drag',
-                    component: () => import(/* webpackChunkName: "drag" */ '../components/page/DragList.vue'),
-                    meta: { title: '拖拽列表' }
-                },
-                {
-                    // 拖拽Dialog组件
-                    path: '/dialog',
-                    component: () => import(/* webpackChunkName: "dragdialog" */ '../components/page/DragDialog.vue'),
-                    meta: { title: '拖拽弹框' }
-                },
-                {
-                    // 国际化组件
-                    path: '/i18n',
-                    component: () => import(/* webpackChunkName: "i18n" */ '../components/page/I18n.vue'),
-                    meta: { title: '国际化' }
-                },
-                {
-                    // 权限页面
-                    path: '/permission',
-                    component: () => import(/* webpackChunkName: "permission" */ '../components/page/Permission.vue'),
-                    meta: { title: '权限测试', permission: true }
-                },
-                {
-                    path: '/404',
-                    component: () => import(/* webpackChunkName: "404" */ '../components/page/404.vue'),
-                    meta: { title: '404' }
-                },
-                {
-                    path: '/403',
-                    component: () => import(/* webpackChunkName: "403" */ '../components/page/403.vue'),
-                    meta: { title: '403' }
-                },
-                {
-                    path: '/donate',
-                    component: () => import(/* webpackChunkName: "donate" */ '../components/page/Donate.vue'),
-                    meta: { title: '支持作者' }
-                }
-            ]
-        },
-        {
-            path: '/login',
-            component: () => import(/* webpackChunkName: "login" */ '../components/page/Login.vue'),
-            meta: { title: '登录' }
-        },
-        {
-            path: '*',
-            redirect: '/404'
-        }
-    ]
+const router = createRouter({
+    history: createWebHistory(process.env.BASE_URL),
+    routes
 });
+
+router.beforeEach((to, from, next) => {
+    document.title = `${to.meta.title} | vue-manage-system`;
+    const role = localStorage.getItem('ms_username');
+    if (!role && to.path !== '/login') {
+        next('/login');
+    } else if (to.meta.permission) {
+        // 如果是管理员权限则可进入，这里只是简单的模拟管理员权限而已
+        role === 'admin'
+            ? next()
+            : next('/403');
+    } else {
+        next();
+    }
+});
+
+export default router;
