@@ -5,7 +5,9 @@
             <el-form :model="param" :rules="rules" ref="login" label-width="0px" class="ms-content">
                 <el-form-item prop="username">
                     <el-input v-model="param.username" placeholder="username">
-                        <el-button slot="prepend" icon="el-icon-lx-people"></el-button>
+                        <template #prepend>
+                            <el-button icon="el-icon-user"></el-button>
+                        </template>
                     </el-input>
                 </el-form-item>
                 <el-form-item prop="password">
@@ -13,9 +15,11 @@
                         type="password"
                         placeholder="password"
                         v-model="param.password"
-                        @keyup.enter.native="submitForm()"
+                        @keyup.enter="submitForm()"
                     >
-                        <el-button slot="prepend" icon="el-icon-lx-lock"></el-button>
+                        <template #prepend>
+                            <el-button icon="el-icon-lock"></el-button>
+                        </template>
                     </el-input>
                 </el-form-item>
                 <div class="login-btn">
@@ -29,33 +33,39 @@
 
 <script>
 export default {
-    data: function() {
+    data() {
         return {
             param: {
-                username: 'admin',
-                password: '123123',
+                username: "admin",
+                password: "123123"
             },
             rules: {
-                username: [{ required: true, message: '请输入用户名', trigger: 'blur' }],
-                password: [{ required: true, message: '请输入密码', trigger: 'blur' }],
-            },
+                username: [
+                    { required: true, message: "请输入用户名", trigger: "blur" }
+                ],
+                password: [
+                    { required: true, message: "请输入密码", trigger: "blur" }
+                ]
+            }
         };
+    },
+    created() {
+        this.$store.commit("clearTags");
     },
     methods: {
         submitForm() {
             this.$refs.login.validate(valid => {
                 if (valid) {
-                    this.$message.success('登录成功');
-                    localStorage.setItem('ms_username', this.param.username);
-                    this.$router.push('/');
+                    this.$message.success("登录成功");
+                    localStorage.setItem("ms_username", this.param.username);
+                    this.$router.push("/");
                 } else {
-                    this.$message.error('请输入账号和密码');
-                    console.log('error submit!!');
+                    this.$message.error("请输入账号和密码");
                     return false;
                 }
             });
-        },
-    },
+        }
+    }
 };
 </script>
 
@@ -64,7 +74,7 @@ export default {
     position: relative;
     width: 100%;
     height: 100%;
-    background-image: url(../../assets/img/login-bg.jpg);
+    background-image: url(../assets/img/login-bg.jpg);
     background-size: 100%;
 }
 .ms-title {
